Добавила всплывающее окно для иконок и для единиц измерения

+++ PillTracker/Models/Pill.swift
@@ -11,7 +11,7 @@ struct Pill {
     let id: UUID
     let icon: UIImage?
     let name: String
-    let dosage: Int
+    let dosage: Double
     let unit: String
     let howToTake: String
     let times: [(hour: String, minute: String)]
